Fix typing for ajv
